+++ src/clj/backtype/storm/testing.clj
@@ -288,7 +288,8 @@
                            (select-keys component->tasks component-ids)
                            component->tasks)
         task-ids (apply concat (vals component->tasks))
-        taskbeats (.taskbeats state storm-id)
+        assignment (.assignment-info state storm-id nil) 
+        taskbeats (.taskbeats state storm-id (:task->node+port assignment))
         heartbeats (dofor [id task-ids] (get taskbeats id))
         stats (dofor [hb heartbeats] (if hb (stat-key (:stats hb)) 0))]
     (reduce + stats)