+++ test/clj/backtype/storm/nimbus_test.clj
@@ -22,20 +22,15 @@
     (count (reverse-map (:task->node+port assignment)))
     ))
 
-(defn mk-task->node+port+taskbeats [state storm-id]
-  (let [task->node+port (:task->node+port (.assignment-info state storm-id nil))
-        node+port->taskbeats (into {} (for [node+port (set (vals task->node+port))]
-                                        {node+port (atom {})}))]
-    (into {} (for [[task node+port] task->node+port
-                   :let [taskbeats (->> task (get task->node+port) (get node+port->taskbeats))]]
-               {task {:node+port node+port :taskbeats taskbeats}}))))
-
-(defn do-task-heartbeat [cluster storm-id task-id task->node+port+taskbeats]
+(defn do-task-heartbeat [cluster storm-id task-id]
   (let [state (:storm-cluster-state cluster)
-        [node port] (-> task->node+port+taskbeats (get task-id) :node+port)
-        taskbeats (-> task->node+port+taskbeats (get task-id) :taskbeats)]
-    (swap! taskbeats assoc task-id (TaskHeartbeat. (current-time-secs) 10 {}))
-    (.worker-heartbeat! state storm-id node port {:storm-id storm-id :taskbeats @taskbeats})))
+        task->node+port (:task->node+port (.assignment-info state storm-id nil))
+        [node port] (get task->node+port task-id)
+        curr-beat (.get-worker-heartbeat state storm-id node port)
+        taskbeats (:taskbeats curr-beat)]
+    (.worker-heartbeat! state storm-id node port
+      {:storm-id storm-id :taskbeats (assoc taskbeats task-id (TaskHeartbeat. (current-time-secs) 10 {}))}
+      )))
 
 (defn task-assignment [cluster storm-id task-id]
   (let [state (:storm-cluster-state cluster)
@@ -204,9 +199,8 @@
       (bind storm-id3 (get-storm-id state "test3"))
 
       (bind task-id (first (task-ids cluster storm-id3)))
-      (bind task->node+port+taskbeats (mk-task->node+port+taskbeats state storm-id3))
       
-      (do-task-heartbeat cluster storm-id3 task-id task->node+port+taskbeats)
+      (do-task-heartbeat cluster storm-id3 task-id)
 
       (.killTopology (:nimbus cluster) "test3")
       (advance-cluster-time cluster 6)
@@ -245,26 +239,25 @@
       (bind [task-id1 task-id2]  (task-ids cluster storm-id))
       (bind ass1 (task-assignment cluster storm-id task-id1))
       (bind ass2 (task-assignment cluster storm-id task-id2))
-      (bind task->node+port+taskbeats (mk-task->node+port+taskbeats state storm-id))      
       
       (advance-cluster-time cluster 59)
-      (do-task-heartbeat cluster storm-id task-id1 task->node+port+taskbeats)
-      (do-task-heartbeat cluster storm-id task-id2 task->node+port+taskbeats)
+      (do-task-heartbeat cluster storm-id task-id1)
+      (do-task-heartbeat cluster storm-id task-id2)
 
       (advance-cluster-time cluster 13)
       (is (= ass1 (task-assignment cluster storm-id task-id1)))
       (is (= ass2 (task-assignment cluster storm-id task-id2)))
-      (do-task-heartbeat cluster storm-id task-id1 task->node+port+taskbeats)
+      (do-task-heartbeat cluster storm-id task-id1)
 
       (advance-cluster-time cluster 11)
-      (do-task-heartbeat cluster storm-id task-id1 task->node+port+taskbeats)
+      (do-task-heartbeat cluster storm-id task-id1)
       (is (= ass1 (task-assignment cluster storm-id task-id1)))
       (check-consistency cluster "test")
 
       ; have to wait an extra 10 seconds because nimbus may not
       ; resynchronize its heartbeat time till monitor-time secs after
       (advance-cluster-time cluster 11)
-      (do-task-heartbeat cluster storm-id task-id1 task->node+port+taskbeats)
+      (do-task-heartbeat cluster storm-id task-id1)
       (is (= ass1 (task-assignment cluster storm-id task-id1)))
       (check-consistency cluster "test")
       
@@ -285,8 +278,8 @@
       (kill-supervisor cluster active-supervisor)
 
       (doseq [i (range 12)]
-        (do-task-heartbeat cluster storm-id task-id1 task->node+port+taskbeats)
-        (do-task-heartbeat cluster storm-id task-id2 task->node+port+taskbeats)
+        (do-task-heartbeat cluster storm-id task-id1)
+        (do-task-heartbeat cluster storm-id task-id2)
         (advance-cluster-time cluster 10)
         )
       ;; tests that it doesn't reassign tasks if they're heartbeating even if supervisor times out