+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
  * Eliminate explicit storage of task->component in Zookeeper
+ * Use worker heartbeats instead of task heartbeats (thanks xuming)
 
 ## 0.7.2 (unreleased but release candidate available)